requirements.txt
# ============================================================
# Smart Inspection Vision System — Requirements
# ============================================================

# --- Core Framework ---
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.3
pydantic-settings==2.2.1

# --- Computer Vision & Deep Learning ---
torch==2.2.1
torchvision==0.17.1
ultralytics==8.1.28          # YOLOv8
opencv-python-headless==4.9.0.80
Pillow==10.2.0
scikit-image==0.22.0         # SSIM for before/after comparison
numpy==1.26.4
albumentations==1.4.0        # Data augmentation

# --- Model Export & Optimization ---
onnx==1.16.0
onnxruntime==1.17.1
# tflite-runtime           # Install separately on target device

# --- MLOps & Experiment Tracking ---
mlflow==2.11.1
optuna==3.5.0                # Hyperparameter tuning

# --- Database ---
sqlalchemy==2.0.28
alembic==1.13.1
aiosqlite==0.20.0            # Async SQLite (dev)
asyncpg==0.29.0              # Async PostgreSQL (prod)

# --- Auth & Security ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# --- Logging & Monitoring ---
loguru==0.7.2
prometheus-fastapi-instrumentator==6.1.0

# --- Testing ---
pytest==8.1.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
httpx==0.27.0                # Async test client

# --- Utilities ---
python-dotenv==1.0.1
PyYAML==6.0.1
tenacity==8.2.3              # Retry logic
aiofiles==23.2.1

# --- Dev Tools ---
black==24.3.0
isort==5.13.2
flake8==7.0.0
mypy==1.9.0
